requirements.txt
# --- Backend (FastAPI & Server) ---
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# --- Database & ORM (Supabase) ---
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# --- Data Science & AI Models ---
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0

# --- Frontend (Streamlit Dashboard) ---
streamlit>=1.25.0
plotly>=5.15.0
requests>=2.31.0

# --- Utility & Configuration ---
python-dotenv>=1.0.0

# --- Testing ---
pytest>=7.4.0
httpx>=0.24.0